requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-import-export==3.3.7
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-spectacular==0.27.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupPy==1.14
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
